+++ common/src/main/kotlin/com/cobblemon/mod/common/battles/interpreter/instructions/InitializeInstruction.kt
@@ -38,7 +38,7 @@ class InitializeInstruction(val instructionSet: InstructionSet, val message: Bat
             pokemon.entity?.let {
                 // If a Pokémon entity is being recalled with an animation,
                 // wrap up the animation and recall the Pokémon immediately.
-                if (it.beamMode == 3) {
+                if (it.beamMode == 3 && battle.battlePartyStores.isEmpty()) {
                     pokemon.effectedPokemon.recall()
                 }
                 // If already on the field, initialize for the BattleInitializePacket