+++ common/src/main/kotlin/com/cobblemon/mod/common/battles/interpreter/instructions/SwitchInstruction.kt
@@ -25,6 +25,8 @@ import com.cobblemon.mod.common.net.serverhandling.storage.SendOutPokemonHandler
 import com.cobblemon.mod.common.net.serverhandling.storage.SendOutPokemonHandler.SEND_OUT_STAGGER_BASE_DURATION
 import com.cobblemon.mod.common.net.serverhandling.storage.SendOutPokemonHandler.SEND_OUT_STAGGER_RANDOM_MAX_DURATION
 import com.cobblemon.mod.common.util.battleLang
+import com.cobblemon.mod.common.util.getPlayer
+import com.cobblemon.mod.common.util.party
 import com.cobblemon.mod.common.util.swap
 import java.util.concurrent.CompletableFuture
 import kotlin.random.Random
@@ -88,6 +90,28 @@ class SwitchInstruction(val instructionSet: InstructionSet, val battleActor: Bat
                     }
                 }
                 GO
+            }
+
+            val futureSwitches = instructionSet.getSubsequentInstructions(this).filterIsInstance<SwitchInstruction>()
+            if (futureSwitches.isEmpty()) {
+                if (battle.battlePartyStores.isNotEmpty()) {
+                    // means battle is using clone teams, recall the "real" pokemon before the sendouts occur
+                    var waitOnRecall = false
+                    battle.actors.forEach { it ->
+                        val playerUUIDS = it.getPlayerUUIDs()
+                        playerUUIDS.forEach { uuid ->
+                            uuid.getPlayer()?.party()?.forEach { pokemon ->
+                                if (pokemon.entity != null) {
+                                    waitOnRecall = true
+                                    pokemon.entity!!.recallWithAnimation()
+                                }
+                            }
+                        }
+                    }
+                    if (waitOnRecall) {
+                        battle.dispatchWaitingToFront(SEND_OUT_DURATION) {  }
+                    }
+                }
             }
         }
         else {